requirements.txt
flask>=2.3
pdfplumber>=0.11
python-docx>=1.1
pyinstaller>=6.0
